+++ src/unitPriceOffer.ts
@@ -3,7 +3,7 @@ import { ItemAvailability } from './itemAvailability';
 import { IAddOn, IName, IOffer } from './offer';
 import { OfferType } from './offerType';
 import { IAmount as IPermitAmount, IDepositAmount, IPaymentAmount } from './permit';
-import { IAppliesToMovieTicket, IPriceSpecification as IUnitPriceSpecification } from './priceSpecification/unitPriceSpecification';
+import { IPriceSpecification as IUnitPriceSpecification } from './priceSpecification/unitPriceSpecification';
 import { IProduct, ProductType } from './product';
 import { IQuantitativeValue } from './quantitativeValue';
 import { SortType } from './sortType';
@@ -49,11 +49,18 @@ export interface IItemOffered {
 }
 /**
  * 単価オファーの価格仕様
+ * 最適化(Pickで表現)(2023-09-01~)
  */
-export type IUnitPriceOfferPriceSpecification = Omit<IUnitPriceSpecification, 'appliesToMovieTicket' | 'project'> & {
+export type IUnitPriceOfferPriceSpecification = Pick<
+    IUnitPriceSpecification,
+    'accounting'
     // Arrayに限定(2022-09-09~)
-    appliesToMovieTicket?: IAppliesToMovieTicket[];
-};
+    | 'appliesToMovieTicket'
+    | 'eligibleQuantity' | 'eligibleTransactionVolume'
+    | 'name' | 'price'
+    | 'priceCurrency' | 'referenceQuantity' | 'typeOf' | 'valueAddedTaxIncluded'
+>;
+// export type IUnitPriceOfferPriceSpecification = Omit<IUnitPriceSpecification, 'appliesToMovieTicket' | 'project'> & {
 export type IAddOnItemOffered = Pick<IProduct, 'typeOf' | 'id' | 'name'>;
 export interface IAddOn4unitPriceOffer extends Pick<IAddOn, 'typeOf' | 'priceCurrency'> {
     itemOffered: IAddOnItemOffered;
@@ -70,17 +77,16 @@ export type IAvailability = ItemAvailability.InStock | ItemAvailability.OutOfSto
 export interface IUnitPriceOffer extends Pick<
     IOffer,
     'project' | 'typeOf' | 'priceCurrency' | 'id' | 'identifier' | 'name' | 'description'
-    | 'alternateName' | 'availability' | 'availableAtOrFrom' | 'itemOffered' | 'priceSpecification'
-    | 'additionalProperty' | 'color' | 'category'
+    | 'alternateName' | 'availability' | 'availableAtOrFrom' | 'itemOffered'
+    | 'priceSpecification' | 'additionalProperty' | 'color' | 'category'
     | 'eligibleSeatingType' | 'eligibleMembershipType' | 'eligibleMonetaryAmount' | 'eligibleSubReservation'
     | 'validFrom' | 'validThrough' | 'validRateLimit'
 > {
-    // advanceBookingRequirementを追加(2023-08-10~)
     /**
      * The amount of time that is required between accepting the offer and the actual usage of the resource or service.
      * 事前予約要件(興行オファー承認日時とイベント開始日時の差)
      */
-    advanceBookingRequirement?: IAdvanceBookingRequirement;
+    advanceBookingRequirement?: IAdvanceBookingRequirement; // 追加(2023-08-10~)
     availability: IAvailability;
     /**
      * コード
@@ -102,6 +108,7 @@ export interface IUnitPriceOffer extends Pick<
  * ソート条件
  */
 export interface ISortOrder {
+    identifier?: SortType;
     'priceSpecification.price'?: SortType;
 }
 /**
@@ -161,7 +168,10 @@ export interface ISearchConditions {
             /**
              * アドオンプロダクトID
              */
-            id?: { $eq?: string };
+            id?: {
+                $eq?: string;
+                $in?: string[];
+            };
         };
     };
     availability?: { $eq?: IAvailability };
@@ -247,4 +257,16 @@ export interface ISearchConditions {
      * 有効期間設定がない、あるいは、有効期間内
      */
     onlyValid?: boolean;
+    parentOffer?: {
+        /**
+         * 集計オファーID
+         */
+        id?: { $in?: string[] };
+    };
+    includedInDataCatalog?: {
+        /**
+         * 記載カタログID
+         */
+        id?: { $in?: string[] };
+    };
 }